fix: Must always cleanup all created resources

+++ vhdbuilder/packer/cleanup.sh
@@ -1,32 +1,16 @@
 #!/bin/bash -x
 
-required_env_vars=(
-  "CLIENT_ID"
-  "CLIENT_SECRET"
-  "TENANT_ID"
-  "SUBSCRIPTION_ID"
-  "PKR_RG_NAME"
-  "MODE"
-  "AZURE_RESOURCE_GROUP_NAME"
-  "SA_NAME"
-)
-
-for v in "${required_env_vars[@]}"; do
-  if [ -z "${!v}" ]; then
-    echo "$v was not set!"
-    exit 1
-  fi
-done
-
 if [[ -z "$SIG_GALLERY_NAME" ]]; then
   SIG_GALLERY_NAME="PackerSigGalleryEastUS"
 fi
 
-#clean up the packer generated resource group
-id=$(az group show --name ${PKR_RG_NAME} | jq .id)
-if [ -n "$id" ]; then
-  echo "Deleting packer resource group ${PKR_RG_NAME}"
-  az group delete --name ${PKR_RG_NAME} --yes
+if [[ -n "$PKR_RG_NAME" ]]; then
+  #clean up the packer generated resource group
+  id=$(az group show --name ${PKR_RG_NAME} | jq .id)
+  if [ -n "$id" ]; then
+    echo "Deleting packer resource group ${PKR_RG_NAME}"
+    az group delete --name ${PKR_RG_NAME} --yes
+  fi
 fi
 
 #clean up the vnet resource group for Windows
@@ -39,10 +23,12 @@ if [ -n "${VNET_RESOURCE_GROUP_NAME}" ]; then
 fi
 
 #clean up managed image
-if [[ "$MODE" != "default" ]]; then
-  id=$(az image show -n ${IMAGE_NAME} -g ${AZURE_RESOURCE_GROUP_NAME} | jq .id)
-  if [ -n "$id" ]; then
-    az image delete -n ${IMAGE_NAME} -g ${AZURE_RESOURCE_GROUP_NAME}
+if [[ -n "$AZURE_RESOURCE_GROUP_NAME" && -n "$IMAGE_NAME" ]]; then
+  if [[ "$MODE" != "default" ]]; then
+    id=$(az image show -n ${IMAGE_NAME} -g ${AZURE_RESOURCE_GROUP_NAME} | jq .id)
+    if [ -n "$id" ]; then
+      az image delete -n ${IMAGE_NAME} -g ${AZURE_RESOURCE_GROUP_NAME}
+    fi
   fi
 fi
 
@@ -82,29 +68,29 @@ if [ ${ARCHITECTURE,,} == "arm64" ] && [ -n "${ARM64_OS_DISK_SNAPSHOT_NAME}" ]; 
 fi
 
 #clean up the temporary storage account
-id=$(az storage account show -n ${SA_NAME} -g ${AZURE_RESOURCE_GROUP_NAME} | jq .id)
-if [ -n "$id" ]; then
-  az storage account delete -n ${SA_NAME} -g ${AZURE_RESOURCE_GROUP_NAME} --yes
+if [[ -n "${SA_NAME}" ]]; then
+  id=$(az storage account show -n ${SA_NAME} -g ${AZURE_RESOURCE_GROUP_NAME} | jq .id)
+  if [ -n "$id" ]; then
+    az storage account delete -n ${SA_NAME} -g ${AZURE_RESOURCE_GROUP_NAME} --yes
+  fi
 fi
 
 #clean up storage account created over a week ago
-EXPIRATION_IN_HOURS=168
-# convert to seconds so we can compare it against the "tags.now" property in the resource group metadata
-(( expirationInSecs = ${EXPIRATION_IN_HOURS} * 60 * 60 ))
-# deadline = the "date +%s" representation of the oldest age we're willing to keep
-(( deadline=$(date +%s)-${expirationInSecs%.*} ))
-echo "Current time is $(date)"
-echo "Looking for storage accounts in ${AZURE_RESOURCE_GROUP_NAME} created over ${EXPIRATION_IN_HOURS} hours ago..."
-echo "That is, those created before $(date -d@$deadline) As shown below"
-az storage account list -g ${AZURE_RESOURCE_GROUP_NAME} | jq --arg dl $deadline '.[] | select(.tags.now < $dl).name' | tr -d '\"' || ""
-for storage_account in $(az storage account list -g ${AZURE_RESOURCE_GROUP_NAME} | jq --arg dl $deadline '.[] | select(.tags.now < $dl).name' | tr -d '\"' || ""); do
-    if [[ "${DRY_RUN}" == "False" ]]; then
-       if [[ $storage_account = aksimages* ]]; then
+if [[ -n "${AZURE_RESOURCE_GROUP_NAME}" && "${DRY_RUN}" == "False" ]]; then
+  EXPIRATION_IN_HOURS=168
+  # convert to seconds so we can compare it against the "tags.now" property in the resource group metadata
+  (( expirationInSecs = ${EXPIRATION_IN_HOURS} * 60 * 60 ))
+  # deadline = the "date +%s" representation of the oldest age we're willing to keep
+  (( deadline=$(date +%s)-${expirationInSecs%.*} ))
+  echo "Current time is $(date)"
+  echo "Looking for storage accounts in ${AZURE_RESOURCE_GROUP_NAME} created over ${EXPIRATION_IN_HOURS} hours ago..."
+  echo "That is, those created before $(date -d@$deadline) As shown below"
+  az storage account list -g ${AZURE_RESOURCE_GROUP_NAME} | jq --arg dl $deadline '.[] | select(.tags.now < $dl).name' | tr -d '\"' || ""
+  for storage_account in $(az storage account list -g ${AZURE_RESOURCE_GROUP_NAME} | jq --arg dl $deadline '.[] | select(.tags.now < $dl).name' | tr -d '\"' || ""); do
+      if [[ $storage_account = aksimages* ]]; then
           echo "Will delete storage account ${storage_account}# from resource group ${AZURE_RESOURCE_GROUP_NAME}..."
           az storage account delete --name ${storage_account} -g ${AZURE_RESOURCE_GROUP_NAME} --yes  || echo "unable to delete storage account ${storage_account}, will continue..."
           echo "Deletion completed"
-        fi
-    else
-        echo "skipping because DRY_RUN is set to True"
-    fi
-done
+      fi
+  done
+fi
